Read simulation inputs from command line args
This commit adds command line arguments for specifying files in which
the Hamiltonian, transition dipole moments, and simulation configuration
parameters are stored. Most notably, simulation parameters were chosen
to be read from a file rather than command line arguments so that there
is a record of the simulation parameters can be saved along side the
simulation inputs and outputs.